fix(android): no network reads on the UI thread; crash/ANR reports (0.1.1)
The API client resumed on the caller's (main) thread after the HTTP call and read and
decoded the response body there. On Android that throws NetworkOnMainThreadException for
any response that isn't already buffered: login worked, chat lists and messages failed,
and uncaught cases crashed the app. Requests now run entirely on Dispatchers.IO;
MainThreadNetworkTest reproduces the Android rule on the JVM (fails on the old code).

Also: cookie store loaded lazily off the main thread and never throws on OkHttp threads;
coroutine safety net + local crash/ANR reports with a Copy button (no message content);
positive scroll offsets for jump-to-message; de-duplicated list keys; decimal media
dimensions; R8 disabled for release until minified builds can be checked on a device.

+++ android_client/core/src/main/kotlin/io/github/revocx35/walogger/core/Models.kt
@@ -174,8 +174,9 @@ data class MediaInfo(
     val mime: String? = null,
     val filename: String? = null,
     val size: Long? = null,
-    val width: Int? = null,
-    val height: Int? = null,
+    // WhatsApp's media dimensions are occasionally non-integers; accept any number.
+    val width: Double? = null,
+    val height: Double? = null,
     val durationSec: Double? = null,
     val thumbDataUrl: String? = null,
 )
